+++ netqasm/__init__.py
@@ -1,5 +1,5 @@
 # Version of the package
-__version__ = "0.0.11"
+__version__ = "0.0.12"
 
 # Version of the language
 NETQASM_VERSION = (0, 0)