+++ netqasm/quantum_gates.py
@@ -40,6 +40,19 @@ def get_rotation_matrix(axis, angle):
     return rot_mat
 
 
+def get_controlled_rotation_matrix(axis, angle) -> np.array:
+    target_pos = get_rotation_matrix(axis, angle)
+    target_neg = get_rotation_matrix(axis, -angle)
+
+    ctrl_zero = np.array([[1, 0], [0, 0]])
+    ctrl_one = np.array([[0, 0], [0, 1]])
+
+    inv_controlled_gate = np.kron(ctrl_zero, target_pos) + np.kron(ctrl_one, np.eye(2))
+    controlled_gate = np.kron(ctrl_one, target_neg) + np.kron(ctrl_zero, np.eye(2))
+
+    return inv_controlled_gate @ controlled_gate
+
+
 def gate_to_matrix(instr, angle=None):
     """Returns the matrix representation of a quantum gate"""
     if instr in STATIC_QUBIT_GATE_TO_MATRIX: